Verificacao ao eliminar vcard pelo admin

+++ backend/app/Http/Controllers/api/VCardController.php
@@ -141,7 +141,6 @@ class VCardController extends Controller
         if ($vcard->transactions()->count() > 0) {
 
             Transaction::where('vcard', $vcard->phone_number)->delete();
-            Category::where('vcard', $vcard->phone_number)->delete();
 
             // delete vcard
             $vcard->delete();
@@ -220,8 +219,36 @@ class VCardController extends Controller
 
     public function destroy(VCard $vcard)
     {
-        $vcard->delete();
-        return new VCardResource($vcard);
+        if ($vcard->balance > 0) {
+            return response()->json(['message' => 'VCard balance must be 0 to be dismissed'], 422);
+        }
+
+        // Delete photo file from storage
+        if ($vcard->photo_url) {
+            if (Storage::exists('public/fotos/' . $vcard->photo_url)) {
+                Storage::delete('public/fotos/' . $vcard->photo_url);
+            }
+        }
+
+        // se tiver transações fazemos um soft delete, se não tiver fazemos um hard delete
+        if ($vcard->transactions()->count() > 0) {
+
+            Transaction::where('vcard', $vcard->phone_number)->delete();
+
+            // delete vcard
+            $vcard->delete();
+        } else {
+            // delete categories
+            // foreach ($vcard->categories() as $category) {
+            //     $category->forceDelete();
+            // }
+            Category::where('vcard', $vcard->phone_number)->forceDelete();
+
+            // delete vcard
+            $vcard->forceDelete();
+        }
+
+        return response()->json(['message' => 'VCard dismissed successfully']);
     }
 
     public function isPhoneNumberAlreadyUsed(Request $request)
